Don't slide on channel dimension when calculate ssim
This affects how ssim statistics are calculated for RGB images, the
previous behavir is somehow "incorrect" that the sliding window for
2d RGB images is a 3d volumn instead of a 2d matrix.

This commit "fixes" this behavior and make it compatible with the
original implementation.

Another internal WIP change that introduced by this commit, is a new
keyword argument `crop` that can be used by MS-SSIM. This also makes
it possible to produce compatible SSIM results to other implementations.